backend/requirements.txt
# Core FastAPI dependencies (Python 3.12 compatible)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# AI & LangGraph (Compatible versions)
langgraph>=0.2.0
langsmith>=0.1.0
google-generativeai>=0.8.0

# File processing
python-multipart>=0.0.6
aiofiles>=23.0.0
pypdf2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
pytest-cov>=4.0.0

# Development tools
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
pre-commit>=3.6.0

# Production
structlog>=24.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Let pip resolve compatible versions
typing-extensions>=4.12.0
